Fix incompability with #7247
See: https://github.com/Automattic/wp-calypso/pull/7451#discussion-diff-75128569

`startEditingPostCopy` temporarily moved in `post-editor/controller.js` as it now contains both Flux and Redux actions to fix an incompability with #7247 regarding the Category Selector (now `EditorTermSelector`), that now uses `post.terms` retrieved from the Redux state.
